+++ controllers/auth/login.js
@@ -1,4 +1,4 @@
-const  User  = require('../../models/user/user');
+const  User  = require('../../models/user');
 const bcrypt =require('bcrypt');
 const jwt = require('jsonwebtoken');
 const { ctrlWrapper, HttpError} = require('../../helpers');
