+++ controllers/auth/logout.js
@@ -1,4 +1,4 @@
-const  User  = require('../../models/user/user');
+const  User  = require('../../models/user');
 const { ctrlWrapper } = require('../../helpers');
