+++ middlewares/authenticate.js
@@ -1,6 +1,6 @@
 const jwt =require('jsonwebtoken');
 const {HttpError} = require('../helpers');
-const User = require('../models/user/user');
+const User = require('../models/user');
 const {SECRET_KEY} = process.env;
 
 const authenticate = async (req, res,next) => {
