+++ models/user.js
@@ -1,6 +1,6 @@
 const {Schema, model} = require('mongoose');
 
-const  handleMongooseError  = require('../../middlewares/handleMongooseError');
+const  handleMongooseError  = require('../middlewares/handleMongooseError');
 
 const emailRegexp = /^[A-Z0-9._%+-]+@[A-Z0-9-]+.+.[A-Z]{2,4}$/i;
 
@@ -30,9 +30,5 @@ const userSchema = new Schema({
 
 userSchema.post('save', handleMongooseError);
 
-
-
 module.exports = model("user", userSchema);
 
-// const User = model('user', userSchema);
-// module.exports = User;
