+++ routes/api/contacts.js
@@ -1,10 +1,9 @@
 const express = require('express');
 const ctrl = require('../../controllers/contacts');
-const validateBody = require('../../middlewares/validateBody')
-const isValidId = require('../../middlewares/isValidId')
-const authenticate = require('../../middlewares/authenticate')
-
-const { schemas } = require('../../models/contact');
+const validateBody = require('../../middlewares/validateBody');
+const isValidId = require('../../middlewares/isValidId');
+const authenticate = require('../../middlewares/authenticate');
+const schemas  = require('../../schemas/contactJoiSchema');
 
 const router = express.Router();
